Refactored response code.
Response is still written across events, but finalisation of response
is now always after other functions writing to response object have
declared that they are finished with it.

A new module has been written to automate this pattern but it needs to
go in its own file, get a better name, and get unit tests.